extracted interface and abstrac class for player controller
