+++ Assets/Scripts/IPlayerController.cs
@@ -1,0 +1,4 @@
+﻿public interface IPlayerController
+{
+    void DoPowerup(float value);
+}
