+++ Assets/Scripts/PlayerController.cs
@@ -1,24 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
-[RequireComponent(typeof( CharacterController ))]
-public class PlayerController : MonoBehaviour
+public abstract class PlayerController : MonoBehaviour, IPlayerController
 {
-    [SerializeField] private float playerSpeedX = 1f;
-    [SerializeField] private float playerSpeedY = 1f;
-
-    private CharacterController characterController;
-
-    void Start()
+    public void DoPowerup(float value)
     {
-        characterController = GetComponent<CharacterController>();
+        throw new System.NotImplementedException();
     }
-    
-    void Update()
-    {
-        var move = new Vector3(Input.GetAxis("Horizontal") * playerSpeedX, Input.GetAxis("Vertical") * playerSpeedY, 0);
-        characterController.Move(move * Time.deltaTime);
-    }
-    
 }
