+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 1d5afd2ac0ffeb84481cdd3c96173c64
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 285b51f08aa84e43b28a787f99f5ff80
+timeCreated: 1663454933
